Move CoordinateMatch to lib
